+++ src/ServiceProvider/ConfigServiceProvider.php
@@ -5,34 +5,15 @@ namespace OctoLab\Cilex\ServiceProvider;
 use Cilex\Application;
 use Cilex\ServiceProviderInterface;
 use OctoLab\Cilex\Command\Config\DumpCommand;
-use OctoLab\Common\Config;
-use Symfony\Component\Config\FileLocator;
 
 /**
  * @author Kamil Samigullin <kamil@samigullin.info>
  */
-class ConfigServiceProvider implements ServiceProviderInterface
+class ConfigServiceProvider
+    extends \OctoLab\Kilex\ServiceProvider\ConfigServiceProvider
+    implements ServiceProviderInterface
 {
-    /** @var string */
-    private $filename;
-    /** @var array */
-    private $placeholders;
-
     /**
-     * @param string $filename
-     * @param array $placeholders
-     *
-     * @api
-     */
-    public function __construct($filename, array $placeholders = [])
-    {
-        $this->filename = $filename;
-        $this->placeholders = $placeholders;
-    }
-
-    /**
-     * @quality:method [B]
-     *
      * @param Application $app
      *
      * @throws \InvalidArgumentException
@@ -46,25 +27,7 @@ class ConfigServiceProvider implements ServiceProviderInterface
      */
     public function register(Application $app)
     {
-        $app['config'] = $app::share(function () {
-            $ext = strtolower(pathinfo($this->filename, PATHINFO_EXTENSION));
-            switch ($ext) {
-                case 'yml':
-                case 'json':
-                    $parser = $ext === 'yml'
-                        ? new Config\Loader\Parser\YamlParser()
-                        : new Config\Loader\Parser\JsonParser();
-                    $loader = new Config\Loader\FileLoader(new FileLocator(), $parser);
-                    $config = (new Config\FileConfig($loader))->load($this->filename, $this->placeholders);
-                    break;
-                case 'php':
-                    $config = (new Config\SimpleConfig(require $this->filename, $this->placeholders));
-                    break;
-                default:
-                    throw new \DomainException(sprintf('File "%s" is not supported.', $this->filename));
-            }
-            return $config;
-        });
+        parent::init($app);
         $app->command(new DumpCommand('config'));
     }
 }