+++ tests/ServiceProvider/ConfigServiceProviderTest.php
@@ -2,7 +2,7 @@
 
 namespace OctoLab\Cilex\ServiceProvider;
 
-use Cilex\Application as CilexApplication;
+use Cilex\Application;
 use OctoLab\Cilex\TestCase;
 
 /**
@@ -10,75 +10,16 @@ use OctoLab\Cilex\TestCase;
  */
 class ConfigServiceProviderTest extends TestCase
 {
-    /** @var array */
-    private $expected = [
-        'app' => [
-            'placeholder_parameter' => 'test',
-            'constant' => E_ALL,
-        ],
-        'component' => [
-            'parameter' => 'base component\'s parameter will be overwritten by root config',
-            'base_parameter' => 'base parameter will not be overwritten',
-        ],
-    ];
-
     /**
      * @test
      * @dataProvider applicationProvider
      *
-     * @param CilexApplication $app
+     * @data Application $app
      */
-    public function registerJson(CilexApplication $app)
+    public function register(Application $app)
     {
-        $app->register($this->getConfigServiceProvider('config', 'json'));
-        foreach ($this->expected as $key => $value) {
-            self::assertEquals($value, $app['config'][$key]);
-        }
-    }
-
-    /**
-     * @test
-     * @dataProvider applicationProvider
-     *
-     * @param CilexApplication $app
-     */
-    public function registerPhp(CilexApplication $app)
-    {
-        $app->register($this->getConfigServiceProvider('config', 'php'));
-        foreach ($this->expected as $key => $value) {
-            self::assertEquals($value, $app['config'][$key]);
-        }
-    }
-
-    /**
-     * @test
-     * @dataProvider applicationProvider
-     *
-     * @param CilexApplication $app
-     */
-    public function registerYaml(CilexApplication $app)
-    {
-        $app->register($this->getConfigServiceProvider('config', 'yml'));
-        foreach ($this->expected as $key => $value) {
-            self::assertEquals($value, $app['config'][$key]);
-        }
-    }
-
-
-    /**
-     * @test
-     * @dataProvider applicationProvider
-     *
-     * @param CilexApplication $app
-     */
-    public function registerUnsupported(CilexApplication $app)
-    {
-        try {
-            $app->register($this->getConfigServiceProvider('config', 'xml'));
-            $app->offsetGet('config');
-            self::fail(sprintf('%s exception expected.', \DomainException::class));
-        } catch (\DomainException $e) {
-            self::assertTrue(true);
-        }
+        $app->register($this->getConfigServiceProvider());
+        self::assertTrue($app->offsetExists('config'));
+        self::assertTrue($app->offsetGet('console')->has('config:dump'));
     }
 }