+++ tests/ServiceProvider/MonologServiceProviderTest.php
@@ -36,25 +36,8 @@ class MonologServiceProviderTest extends TestCase
      *
      * @param Application $app
      */
-    public function registerFailure(Application $app)
-    {
-        try {
-            $app->register(new MonologServiceProvider());
-            self::fail(sprintf('%s exception expected.', \InvalidArgumentException::class));
-        } catch (\InvalidArgumentException $e) {
-            self::assertTrue(true);
-        }
-    }
-
-    /**
-     * @test
-     * @dataProvider applicationProvider
-     *
-     * @param Application $app
-     */
     public function registerEmpty(Application $app)
     {
-        $app->register($this->getConfigServiceProvider('config', 'yml'));
         $app->register(new MonologServiceProvider());
         try {
             $app->offsetGet('logger');