+++ tests/TestCase.php
@@ -23,14 +23,13 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase
 
     /**
      * @param string $config
-     * @param string $extension
      * @param array $placeholders
      *
      * @return ServiceProvider\ConfigServiceProvider
      */
-    protected function getConfigServiceProvider($config, $extension, array $placeholders = ['placeholder' => 'test'])
+    protected function getConfigServiceProvider($config = 'config', array $placeholders = ['placeholder' => 'test'])
     {
-        return new ServiceProvider\ConfigServiceProvider($this->getConfigPath($config, $extension), $placeholders);
+        return new ServiceProvider\ConfigServiceProvider($this->getConfigPath($config), $placeholders);
     }
 
     /**
@@ -38,7 +37,7 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase
      */
     protected function getConfigServiceProviderForMonolog()
     {
-        return $this->getConfigServiceProvider('monolog/config', 'yml', ['root_dir' => __DIR__]);
+        return $this->getConfigServiceProvider('monolog/config', ['root_dir' => __DIR__]);
     }
 
     /**
@@ -46,7 +45,7 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase
      */
     protected function getConfigServiceProviderForDoctrine()
     {
-        return $this->getConfigServiceProvider('doctrine/config', 'yml');
+        return $this->getConfigServiceProvider('doctrine/config');
     }
 
     /**
@@ -54,18 +53,17 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase
      */
     protected function getConfigServiceProviderForCliMenu()
     {
-        return $this->getConfigServiceProvider('cli-menu/config', 'yml');
+        return $this->getConfigServiceProvider('cli-menu/config');
     }
 
     /**
      * @param string $config
-     * @param string $extension
      *
      * @return string
      */
-    protected function getConfigPath($config = 'config', $extension = 'yml')
+    protected function getConfigPath($config = 'config')
     {
-        return sprintf('%s/app/config/%s.%s', __DIR__, $config, $extension);
+        return sprintf('%s/app/config/%s.yml', __DIR__, $config);
     }
 
     /**