+++ tests/app/config/component/base.php
@@ -1,7 +1,0 @@
-<?php
-return [
-    'component' => [
-        'parameter' => 'base parameter will be overwritten',
-        'base_parameter' => 'base parameter will not be overwritten',
-    ],
-];