+++ tests/app/config/parameters.php
@@ -1,6 +1,0 @@
-<?php
-return [
-    'parameters' => [
-        'parameter' => 'will overwrite parameter',
-    ],
-];